+++ src/app/modules/mytodosSlice.jsx
@@ -24,19 +24,6 @@ const initialState = {
 //   }
 // );
 
-// 나의 ToDo 피드를 조회할 때 사용되는 thunk action creater 
-export const getSetUpMyTodoFetch = createAsyncThunk(
-  'mytodos/getSetUpMyTodoFetch',
-  async (payload, thunkAPI) => {
-    try {
-      const response = await instance.get(`/mytodos?date=${payload.date}`);
-      return thunkAPI.fulfillWithValue(response.data);
-    } catch (error) {
-      return thunkAPI.rejectWithValue(error.response.data);
-    }
-  }
-)
-
 // 타인의 todo 피드 조회
 export const getOthersTodoFetch = createAsyncThunk(
   'otherstodos/getOthersTodoFetch',
@@ -73,22 +60,6 @@ const mytodosSlice = createSlice({
     //   newState.message = action.payload.message;
     //   return newState;
     // })
-
-    //getSetUpMyTodoFetch Creater 작동 시 적용되는 내용들
-    builder.addCase(getSetUpMyTodoFetch.pending , (state, action)=> {
-      return state;
-    })
-    builder.addCase(getSetUpMyTodoFetch.fulfilled, (state, action)=> {
-      const newState ={...state}
-      newState.message = action.payload.message;
-      newState.data = action.payload.data;
-      return newState;
-    })
-    builder.addCase(getSetUpMyTodoFetch.rejected, (state, action)=> {
-      const newState = {...state };
-      newState.errorMessage = action.payload.errorMessage;
-      return newState;
-    })
     
     //타인의 todo
     builder.addCase(getOthersTodoFetch.pending , (state, action)=> {
