+++ src/app/store.js
@@ -4,6 +4,7 @@ import mytodosSlice from './modules/mytodosSlice';
 import todolistsSlice from './modules/todolistsSlice';
 import detailSlice from './modules/detailSlice';
 import followSlice from './modules/followSlice';
+import setUpTodosSlice from './modules/setUpTodoSlice';
 
 export const store = configureStore({
   reducer: {
@@ -12,5 +13,6 @@ export const store = configureStore({
     todolists:todolistsSlice.reducer,
     detail:detailSlice.reducer,
     follow:followSlice.reducer,
+    setuptodos:setUpTodosSlice.reducer,
   },
 });
